% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lmannotate.R
\name{geom_lmannotate}
\alias{geom_lmannotate}
\title{Add an annotation describing a linear model to a ggplot2 plot}
\usage{
geom_lmannotate(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, facet_mode = FALSE, ...)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2}

\item{facet_mode}{If TRUE (FALSE by default), will draw every annotation in
the top right corner regardless of how many groups there are}
}
\description{
Add an annotation describing a linear model to a ggplot2 plot
}
