% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lmannotate.R
\name{geom_lmannotate}
\alias{geom_lmannotate}
\title{Add an annotation describing a linear model to a ggplot2 plot}
\usage{
geom_lmannotate(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, glue_exp = NULL, ...)
}
\arguments{
\item{mapping, data, stat, position, na.rm, show.legend, inherit.aes, ...}{As
standard for ggplot2}

\item{glue_exp}{An expression to be parsed by `glue::glue()`, to form the
annotation describing each linear model. The fitted model object returned by
`lm()` is available as the variable `model`}
}
\description{
Add an annotation describing a linear model to a ggplot2 plot
}
